# lib/api-zod/src/generated/types/decisionIntakeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TerraFuse environmental decision firewall API
 * OpenAPI spec version: 0.1.0
 */

export interface DecisionIntakeInput {
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  sourceText?: string;
}