# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TerraFuse environmental decision firewall API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}